refactor :: 색 리페토링

+++ src/app/api/categories/[id]/route.ts
@@ -26,10 +26,23 @@ export async function PUT(req: NextRequest, { params }: Params) {
 export async function DELETE(_: NextRequest, { params }: Params) {
   try {
     const id = Number(params.id);
+
+
+    const relatedMenus = await prisma.menu.findFirst({ where: { categoryId: id } });
+
+    if (relatedMenus) {
+ 
+      return NextResponse.json(
+        { error: "해당 카테고리에 메뉴가 존재하여 삭제할 수 없습니다." },
+        { status: 400 }
+      );
+    }
+
     await prisma.category.delete({ where: { id } });
 
     return NextResponse.json({ success: true });
   } catch (error) {
+    console.error("Category delete error:", error);
     return NextResponse.json({ error: "카테고리 삭제 실패" }, { status: 500 });
   }
 }
